Remove split between token part and parsing part
