+++ include/parsing.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parsing.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: abasdere <abasdere@student.42.fr>          +#+  +:+       +#+        */
+/*   By: gmoulin <gmoulin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/23 17:49:21 by gmoulin           #+#    #+#             */
-/*   Updated: 2025/01/31 14:19:50 by abasdere         ###   ########.fr       */
+/*   Updated: 2025/01/31 16:06:42 by gmoulin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,42 @@
 # define PARSING_H
 
 # include "data_struct.h"
+
+# define CHAR_ERROR "`(){}[];!\\*"
+
+typedef enum s_token_type
+{
+	PIPE,
+	REDIR_IN,
+	REDIR_OUT,
+	STRING_SQ,
+	STRING_DQ,
+	WORD,
+}	t_token_type;
+
+typedef struct s_token
+{
+	t_token_type	type;
+	char			*value;
+	struct s_token	*next;
+	struct s_token	*prev;
+}	t_token;
+
+//token_utils.c
+t_token		*new_token(t_token_type type, const char *value);
+void		append_token(t_token **head, t_token *new_token);
+void		free_token(t_token *token);
+void		free_token_list(t_token **head);
+
+//token_utils2.c
+void		print_token_list(t_token *head);
+const char	*tttostr(t_token_type type);
+
+//tokenize.c
+int			tokenize(char *input, t_token **head);
+
+int			parse_input(char *user_input, t_tree **tree, const t_env *env);
+
 
 //checks.c
 int		check_open_quotes(const char *rl);
